refactor: split unit test
