+++ src/flow/src/transform.rs
@@ -163,7 +163,7 @@ mod test {
 
     use super::*;
 
-    fn create_test_ctx() -> DataflowContext {
+    pub fn create_test_ctx() -> DataflowContext {
         let gid = GlobalId::User(0);
         let name = vec!["numbers".to_string()];
         let schema = RelationType::new(vec![ColumnType::new(CDT::uint32_datatype(), false)]);
@@ -175,7 +175,7 @@ mod test {
         }
     }
 
-    fn create_test_query_engine() -> Arc<dyn QueryEngine> {
+    pub fn create_test_query_engine() -> Arc<dyn QueryEngine> {
         let catalog_list = catalog::memory::new_memory_catalog_manager().unwrap();
         let req = RegisterTableRequest {
             catalog: DEFAULT_CATALOG_NAME.to_string(),
@@ -193,7 +193,7 @@ mod test {
         engine
     }
 
-    async fn sql_to_substrait(engine: Arc<dyn QueryEngine>, sql: &str) -> proto::Plan {
+    pub async fn sql_to_substrait(engine: Arc<dyn QueryEngine>, sql: &str) -> proto::Plan {
         // let engine = create_test_query_engine();
         let stmt = QueryLanguageParser::parse_sql(sql, &QueryContext::arc()).unwrap();
         let plan = engine
@@ -207,359 +207,5 @@ mod test {
         let bytes = DFLogicalSubstraitConvertor {}.encode(&plan).unwrap();
 
         proto::Plan::decode(bytes).unwrap()
-    }
-
-    /// test if literal in substrait plan can be correctly converted to flow plan
-    #[tokio::test]
-    async fn test_literal() {
-        let engine = create_test_query_engine();
-        let sql = "SELECT 1 FROM numbers";
-        let plan = sql_to_substrait(engine.clone(), sql).await;
-
-        let mut ctx = create_test_ctx();
-        let flow_plan = TypedPlan::from_substrait_plan(&mut ctx, &plan);
-
-        let expected = TypedPlan {
-            typ: RelationType::new(vec![ColumnType::new(CDT::int64_datatype(), true)]),
-            plan: Plan::Constant {
-                rows: vec![(
-                    repr::Row::new(vec![Value::Int64(1)]),
-                    repr::Timestamp::MIN,
-                    1,
-                )],
-            },
-        };
-
-        assert_eq!(flow_plan.unwrap(), expected);
-    }
-
-    #[tokio::test]
-    async fn test_where_and() {
-        let engine = create_test_query_engine();
-        let sql = "SELECT number FROM numbers WHERE number >= 1 AND number <= 3 AND number!=2";
-        let plan = sql_to_substrait(engine.clone(), sql).await;
-
-        let mut ctx = create_test_ctx();
-        let flow_plan = TypedPlan::from_substrait_plan(&mut ctx, &plan);
-
-        // TODO(discord9): optimize binary and to variadic and
-        let filter = ScalarExpr::CallVariadic {
-            func: VariadicFunc::And,
-            exprs: vec![
-                ScalarExpr::Column(0).call_binary(
-                    ScalarExpr::Literal(Value::from(1u32), CDT::uint32_datatype()),
-                    BinaryFunc::Gte,
-                ),
-                ScalarExpr::Column(0).call_binary(
-                    ScalarExpr::Literal(Value::from(3u32), CDT::uint32_datatype()),
-                    BinaryFunc::Lte,
-                ),
-                ScalarExpr::Column(0).call_binary(
-                    ScalarExpr::Literal(Value::from(2u32), CDT::uint32_datatype()),
-                    BinaryFunc::NotEq,
-                ),
-            ],
-        };
-        let expected = TypedPlan {
-            typ: RelationType::new(vec![ColumnType::new(CDT::uint32_datatype(), false)]),
-            plan: Plan::Mfp {
-                input: Box::new(Plan::Get {
-                    id: crate::expr::Id::Global(GlobalId::User(0)),
-                }),
-                mfp: MapFilterProject::new(1)
-                    .map(vec![ScalarExpr::Column(0)])
-                    .unwrap()
-                    .filter(vec![filter])
-                    .unwrap()
-                    .project(vec![1])
-                    .unwrap(),
-            },
-        };
-        assert_eq!(flow_plan.unwrap(), expected);
-    }
-
-    /// case: binary functions&constant folding can happen in converting substrait plan
-    #[tokio::test]
-    async fn test_binary_func_and_constant_folding() {
-        let engine = create_test_query_engine();
-        let sql = "SELECT 1+1*2-1/1+1%2==3 FROM numbers";
-        let plan = sql_to_substrait(engine.clone(), sql).await;
-
-        let mut ctx = create_test_ctx();
-        let flow_plan = TypedPlan::from_substrait_plan(&mut ctx, &plan);
-
-        let expected = TypedPlan {
-            typ: RelationType::new(vec![ColumnType::new(CDT::boolean_datatype(), true)]),
-            plan: Plan::Constant {
-                rows: vec![(
-                    repr::Row::new(vec![Value::from(true)]),
-                    repr::Timestamp::MIN,
-                    1,
-                )],
-            },
-        };
-
-        assert_eq!(flow_plan.unwrap(), expected);
-    }
-
-    /// test if the type of the literal is correctly inferred, i.e. in here literal is decoded to be int64, but need to be uint32,
-    #[tokio::test]
-    async fn test_implicitly_cast() {
-        let engine = create_test_query_engine();
-        let sql = "SELECT number+1 FROM numbers";
-        let plan = sql_to_substrait(engine.clone(), sql).await;
-
-        let mut ctx = create_test_ctx();
-        let flow_plan = TypedPlan::from_substrait_plan(&mut ctx, &plan);
-
-        let expected = TypedPlan {
-            typ: RelationType::new(vec![ColumnType::new(CDT::uint32_datatype(), true)]),
-            plan: Plan::Mfp {
-                input: Box::new(Plan::Get {
-                    id: crate::expr::Id::Global(GlobalId::User(0)),
-                }),
-                mfp: MapFilterProject::new(1)
-                    .map(vec![ScalarExpr::Column(0).call_binary(
-                        ScalarExpr::Literal(Value::from(1u32), CDT::uint32_datatype()),
-                        BinaryFunc::AddUInt32,
-                    )])
-                    .unwrap()
-                    .project(vec![1])
-                    .unwrap(),
-            },
-        };
-        assert_eq!(flow_plan.unwrap(), expected);
-    }
-
-    #[tokio::test]
-    async fn test_cast() {
-        let engine = create_test_query_engine();
-        let sql = "SELECT CAST(1 AS INT16) FROM numbers";
-        let plan = sql_to_substrait(engine.clone(), sql).await;
-
-        let mut ctx = create_test_ctx();
-        let flow_plan = TypedPlan::from_substrait_plan(&mut ctx, &plan);
-
-        let expected = TypedPlan {
-            typ: RelationType::new(vec![ColumnType::new(CDT::int16_datatype(), true)]),
-            plan: Plan::Mfp {
-                input: Box::new(Plan::Get {
-                    id: crate::expr::Id::Global(GlobalId::User(0)),
-                }),
-                mfp: MapFilterProject::new(1)
-                    .map(vec![ScalarExpr::Literal(
-                        Value::Int64(1),
-                        CDT::int64_datatype(),
-                    )
-                    .call_unary(UnaryFunc::Cast(CDT::int16_datatype()))])
-                    .unwrap()
-                    .project(vec![1])
-                    .unwrap(),
-            },
-        };
-        assert_eq!(flow_plan.unwrap(), expected);
-    }
-
-    #[tokio::test]
-    async fn test_select() {
-        let engine = create_test_query_engine();
-        let sql = "SELECT number FROM numbers";
-        let plan = sql_to_substrait(engine.clone(), sql).await;
-
-        let mut ctx = create_test_ctx();
-        let flow_plan = TypedPlan::from_substrait_plan(&mut ctx, &plan);
-
-        let expected = TypedPlan {
-            typ: RelationType::new(vec![ColumnType::new(CDT::uint32_datatype(), false)]),
-            plan: Plan::Mfp {
-                input: Box::new(Plan::Get {
-                    id: crate::expr::Id::Global(GlobalId::User(0)),
-                }),
-                mfp: MapFilterProject::new(1)
-                    .map(vec![ScalarExpr::Column(0)])
-                    .unwrap()
-                    .project(vec![1])
-                    .unwrap(),
-            },
-        };
-
-        assert_eq!(flow_plan.unwrap(), expected);
-    }
-
-    #[tokio::test]
-    async fn test_select_add() {
-        let engine = create_test_query_engine();
-        let sql = "SELECT number+number FROM numbers";
-        let plan = sql_to_substrait(engine.clone(), sql).await;
-
-        let mut ctx = create_test_ctx();
-        let flow_plan = TypedPlan::from_substrait_plan(&mut ctx, &plan);
-
-        let expected = TypedPlan {
-            typ: RelationType::new(vec![ColumnType::new(CDT::uint32_datatype(), true)]),
-            plan: Plan::Mfp {
-                input: Box::new(Plan::Get {
-                    id: crate::expr::Id::Global(GlobalId::User(0)),
-                }),
-                mfp: MapFilterProject::new(1)
-                    .map(vec![ScalarExpr::Column(0)
-                        .call_binary(ScalarExpr::Column(0), BinaryFunc::AddUInt32)])
-                    .unwrap()
-                    .project(vec![1])
-                    .unwrap(),
-            },
-        };
-
-        assert_eq!(flow_plan.unwrap(), expected);
-    }
-
-    #[tokio::test]
-    async fn test_sum() {
-        let engine = create_test_query_engine();
-        let sql = "SELECT sum(number) FROM numbers";
-        let plan = sql_to_substrait(engine.clone(), sql).await;
-
-        let mut ctx = create_test_ctx();
-        let flow_plan = TypedPlan::from_substrait_plan(&mut ctx, &plan);
-
-        let aggr_expr = AggregateExpr {
-            func: AggregateFunc::SumUInt32,
-            expr: ScalarExpr::Column(0),
-            distinct: false,
-        };
-        let expected = TypedPlan {
-            typ: RelationType::new(vec![ColumnType::new(CDT::uint32_datatype(), true)]),
-            plan: Plan::Mfp {
-                input: Box::new(Plan::Reduce {
-                    input: Box::new(Plan::Get {
-                        id: crate::expr::Id::Global(GlobalId::User(0)),
-                    }),
-                    key_val_plan: KeyValPlan {
-                        key_plan: MapFilterProject::new(1)
-                            .project(vec![])
-                            .unwrap()
-                            .into_safe(),
-                        val_plan: MapFilterProject::new(1)
-                            .project(vec![0])
-                            .unwrap()
-                            .into_safe(),
-                    },
-                    reduce_plan: ReducePlan::Accumulable(AccumulablePlan {
-                        full_aggrs: vec![aggr_expr.clone()],
-                        simple_aggrs: vec![(0, 0, aggr_expr.clone())],
-                        distinct_aggrs: vec![],
-                    }),
-                }),
-                mfp: MapFilterProject::new(1)
-                    .map(vec![ScalarExpr::Column(0)])
-                    .unwrap()
-                    .project(vec![1])
-                    .unwrap(),
-            },
-        };
-        assert_eq!(flow_plan.unwrap(), expected);
-    }
-
-    #[tokio::test]
-    async fn test_sum_group_by() {
-        let engine = create_test_query_engine();
-        let sql = "SELECT sum(number), number FROM numbers GROUP BY number";
-        let plan = sql_to_substrait(engine.clone(), sql).await;
-
-        let mut ctx = create_test_ctx();
-        let flow_plan = TypedPlan::from_substrait_plan(&mut ctx, &plan).unwrap();
-
-        let aggr_expr = AggregateExpr {
-            func: AggregateFunc::SumUInt32,
-            expr: ScalarExpr::Column(0),
-            distinct: false,
-        };
-        let expected = TypedPlan {
-            typ: RelationType::new(vec![
-                ColumnType::new(CDT::uint32_datatype(), true),
-                ColumnType::new(CDT::uint32_datatype(), false),
-            ]),
-            plan: Plan::Mfp {
-                input: Box::new(Plan::Reduce {
-                    input: Box::new(Plan::Get {
-                        id: crate::expr::Id::Global(GlobalId::User(0)),
-                    }),
-                    key_val_plan: KeyValPlan {
-                        key_plan: MapFilterProject::new(1)
-                            .map(vec![ScalarExpr::Column(0)])
-                            .unwrap()
-                            .project(vec![1])
-                            .unwrap()
-                            .into_safe(),
-                        val_plan: MapFilterProject::new(1)
-                            .project(vec![0])
-                            .unwrap()
-                            .into_safe(),
-                    },
-                    reduce_plan: ReducePlan::Accumulable(AccumulablePlan {
-                        full_aggrs: vec![aggr_expr.clone()],
-                        simple_aggrs: vec![(0, 0, aggr_expr.clone())],
-                        distinct_aggrs: vec![],
-                    }),
-                }),
-                mfp: MapFilterProject::new(2)
-                    .map(vec![ScalarExpr::Column(1), ScalarExpr::Column(0)])
-                    .unwrap()
-                    .project(vec![2, 3])
-                    .unwrap(),
-            },
-        };
-
-        assert_eq!(flow_plan, expected);
-    }
-
-    #[tokio::test]
-    async fn test_sum_add() {
-        let engine = create_test_query_engine();
-        let sql = "SELECT sum(number+number) FROM numbers";
-        let plan = sql_to_substrait(engine.clone(), sql).await;
-
-        let mut ctx = create_test_ctx();
-        let flow_plan = TypedPlan::from_substrait_plan(&mut ctx, &plan);
-
-        let aggr_expr = AggregateExpr {
-            func: AggregateFunc::SumUInt32,
-            expr: ScalarExpr::Column(0),
-            distinct: false,
-        };
-        let expected = TypedPlan {
-            typ: RelationType::new(vec![ColumnType::new(CDT::uint32_datatype(), true)]),
-            plan: Plan::Mfp {
-                input: Box::new(Plan::Reduce {
-                    input: Box::new(Plan::Get {
-                        id: crate::expr::Id::Global(GlobalId::User(0)),
-                    }),
-                    key_val_plan: KeyValPlan {
-                        key_plan: MapFilterProject::new(1)
-                            .project(vec![])
-                            .unwrap()
-                            .into_safe(),
-                        val_plan: MapFilterProject::new(1)
-                            .map(vec![ScalarExpr::Column(0)
-                                .call_binary(ScalarExpr::Column(0), BinaryFunc::AddUInt32)])
-                            .unwrap()
-                            .project(vec![1])
-                            .unwrap()
-                            .into_safe(),
-                    },
-                    reduce_plan: ReducePlan::Accumulable(AccumulablePlan {
-                        full_aggrs: vec![aggr_expr.clone()],
-                        simple_aggrs: vec![(0, 0, aggr_expr.clone())],
-                        distinct_aggrs: vec![],
-                    }),
-                }),
-                mfp: MapFilterProject::new(1)
-                    .map(vec![ScalarExpr::Column(0)])
-                    .unwrap()
-                    .project(vec![1])
-                    .unwrap(),
-            },
-        };
-        assert_eq!(flow_plan.unwrap(), expected);
     }
 }
